Refactored app.rb into rentals handler
